server/requirements.txt
boto3==1.36.1
fastapi==0.116.1
pydantic==2.11.7
PyJWT==2.10.1
bcrypt==4.3.0
SQLAlchemy==2.0.37
psycopg2==2.9.10